Change to ClientV2ConfigurationUtils
